+++ jstests/aggregation/sources/facet/inner_lookup.js
@@ -1,9 +1,3 @@
-// Cannot implicitly shard accessed collections because unsupported use of sharded collection
-// for target collection of $lookup and $graphLookup.
-// @tags: [
-//   assumes_unsharded_collection,
-// ]
-
 /**
  * Tests that using a $lookup stage inside of a $facet stage will yield the same results as using
  * the $lookup stage outside of the $facet stage.
@@ -11,8 +5,19 @@
 (function() {
 "use strict";
 
+load("jstests/libs/fixture_helpers.js");  // For isSharded.
+
 var local = db.facetLookupLocal;
 var foreign = db.facetLookupForeign;
+
+// Do not run the rest of the tests if the foreign collection is implicitly sharded but the flag to
+// allow $lookup/$graphLookup into a sharded collection is disabled.
+const getShardedLookupParam = db.adminCommand({getParameter: 1, featureFlagShardedLookup: 1});
+const isShardedLookupEnabled = getShardedLookupParam.hasOwnProperty("featureFlagShardedLookup") &&
+    getShardedLookupParam.featureFlagShardedLookup.value;
+if (FixtureHelpers.isSharded(foreign) && !isShardedLookupEnabled) {
+    return;
+}
 
 local.drop();
 assert.commandWorked(local.insert({_id: 0}));
